dom/nodes/Node-appendChild-text-and-script-in-style.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting text and script nodes in a style element</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<style id="style"></style>
<script>
const happened = []
const style = document.getElementById("style");
test(() => {
  const r1 = new Text("body {}");
  const r2 = new Text("body {}");
  const script = document.createElement("script");
  script.textContent = `
    happened.push(style.sheet.cssRules.length);
`;

  const df = document.createDocumentFragment();
  df.appendChild(r1);
  df.appendChild(script);
  df.appendChild(r2);

  assert_array_equals(happened, []);
  style.appendChild(df);
  assert_array_equals(happened, [2]);
});
</script>
